+++ src/module/sheets/SwadeBaseActorSheet.ts
@@ -13,6 +13,13 @@ export default class SwadeBaseActorSheet extends ActorSheet {
     // Everything below here is only needed if the sheet is editable
     if (!this.options.editable) return;
 
+    const inputs = html.find('input');
+    inputs.on('focus', (ev) => ev.currentTarget.select());
+    inputs
+      .addBack()
+      .find('[data-dtype="Number"]')
+      .on('change', this._onChangeInputDelta.bind(this));
+
     if (this.actor.isOwner) {
       const handler = (ev: DragEvent) => this._onDragStart(ev);
       html.find('li.active-effect').each((i, li) => {
@@ -221,6 +228,9 @@ export default class SwadeBaseActorSheet extends ActorSheet {
         flavor: statData.label,
       });
     });
+
+    //Wealth Die Roll
+    html.find('.currency .roll').on('click', () => this.actor.rollWealthDie());
   }
 
   getData() {
@@ -281,6 +291,8 @@ export default class SwadeBaseActorSheet extends ActorSheet {
       data.settingrules = {
         conviction: game.settings.get('swade', 'enableConviction'),
         noPowerPoints: game.settings.get('swade', 'noPowerPoints'),
+        wealthType: game.settings.get('swade', 'wealthType'),
+        currencyName: game.settings.get('swade', 'currencyName'),
       };
     }
 
@@ -481,5 +493,20 @@ export default class SwadeBaseActorSheet extends ActorSheet {
         ct.classList.remove('active');
       }
     });
+  }
+
+  /**
+   * Handle input changes to numeric form fields, allowing them to accept delta-typed inputs
+   * @param {Event} event  Triggering event.
+   */
+  protected _onChangeInputDelta(event: Event) {
+    const input = event.target as HTMLInputElement;
+    const value = input.value;
+    if (['+', '-'].includes(value[0])) {
+      const delta = parseInt(value, 10);
+      input.value = getProperty(this.actor.data, input.name) + delta;
+    } else if (value[0] === '=') {
+      input.value = value.slice(1);
+    }
   }
 }