+++ src/module/sheets/official/CharacterSheet.ts
@@ -100,6 +100,7 @@ export default class CharacterSheet extends ActorSheet {
         li.setAttribute('draggable', 'true');
         li.addEventListener('dragstart', handler, false);
       });
+
       html
         .find('.status input[type="checkbox"]')
         .on('change', async (event) => {
@@ -532,24 +533,7 @@ export default class CharacterSheet extends ActorSheet {
     });
 
     //Wealth Die Roll
-    html.find('.currency .roll').on('click', () => {
-      if (this.actor.data.type === 'vehicle') return;
-      const die = this.actor.data.data.details.wealth.die ?? 6;
-      const mod = this.actor.data.data.details.wealth.modifier ?? 0;
-      const wildDie = this.actor.data.data.details.wealth['wild-die'] ?? 6;
-      const dieLabel = game.i18n.localize('SWADE.WealthDie');
-      const wildDieLabel = game.i18n.localize('SWADE.WildDie');
-      const formula = `{1d${die}x[${dieLabel}], 1d${wildDie}x[${wildDieLabel}]}kh`;
-
-      game.swade.RollDialog.asPromise({
-        roll: new Roll(formula),
-        mods: [{ label: 'Modifier', value: mod }],
-        speaker: ChatMessage.getSpeaker(),
-        actor: this.actor,
-        flavor: game.i18n.localize('SWADE.WealthDie'),
-        title: game.i18n.localize('SWADE.WealthDie'),
-      });
-    });
+    html.find('.currency .roll').on('click', () => this.actor.rollWealthDie());
   }
 
   getData() {